Added NM_CONFIG_DIR option to the Makefile (closes #68)
This should not be used in most cases, as it will break compatibility with
existing mods which depend on adding a config to NM. This option is meant as a
better option than find/replace for people who do private builds of NM for
themselves and want to change the config dir.

It will also affect the path used for installing the documentation and shown in
the default menu item.